requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
psycopg2-binary==2.9.5
SQLAlchemy==2.0.7
dependency-injector==4.41.0
alembic==1.10.2
email-validator==1.3.1
python-jose==3.3.0
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
